+++ justfile
@@ -4,7 +4,7 @@ export RUST_BACKTRACE := "full"
 export SKIP_WASM_BUILD := "1"
 export RUST_BIN_DIR := "target/x86_64-unknown-linux-gnu"
 export TARGET := "x86_64-unknown-linux-gnu"
-export RUSTV := "nightly-2024-03-05"
+export RUSTV := "stable"
 export RELEASE_NAME := "development"
 
 fmt:
@@ -38,3 +38,15 @@ fix:
   @echo "Running cargo fix..."
   cargo +{{RUSTV}} fix --workspace
   git diff --exit-code || (echo "There are local changes after running 'cargo fix --workspace' ❌" && exit 1)
+
+lint:
+  @echo "Running cargo fmt..."
+  cargo +{{RUSTV}} fmt --all
+  @echo "Running cargo clippy with automatic fixes on potentially dirty code..."
+  cargo +{{RUSTV}} clippy --fix --allow-dirty -- -A clippy::panic \
+                                                      -A clippy::todo \
+                                                      -A clippy::unimplemented
+  @echo "Running cargo clippy..."
+  cargo +{{RUSTV}} clippy -- -D clippy::panic \
+                            -D clippy::todo \
+                            -D clippy::unimplemented
