+++ pallets/admin-utils/tests/mock.rs
@@ -108,7 +108,9 @@ parameter_types! {
     pub const InitialSubnetLimit: u16 = 10; // Max 10 subnets.
     pub const InitialNetworkRateLimit: u64 = 0;
     pub const InitialTargetStakesPerInterval: u16 = 1;
-
+    pub const InitialAlphaHigh: u16 = 900; // Represents 0.9 as per the production default
+    pub const InitialAlphaLow: u16 = 700; // Represents 0.7 as per the production default
+    pub const InitialLiquidAlphaOn: bool = false; // Default value for LiquidAlphaOn
 }
 
 impl pallet_subtensor::Config for Test {
@@ -160,6 +162,9 @@ impl pallet_subtensor::Config for Test {
     type InitialSubnetLimit = InitialSubnetLimit;
     type InitialNetworkRateLimit = InitialNetworkRateLimit;
     type InitialTargetStakesPerInterval = InitialTargetStakesPerInterval;
+    type AlphaHigh = InitialAlphaHigh;
+    type AlphaLow = InitialAlphaLow;
+    type LiquidAlphaOn = InitialLiquidAlphaOn;
 }
 
 #[derive_impl(frame_system::config_preludes::TestDefaultConfig)]
@@ -461,6 +466,18 @@ impl pallet_admin_utils::SubtensorInterface<AccountId, Balance, RuntimeOrigin> f
 
     fn set_commit_reveal_weights_enabled(netuid: u16, enabled: bool) {
         SubtensorModule::set_commit_reveal_weights_enabled(netuid, enabled);
+    }
+
+    fn set_alpha_high(netuid: u16, alpha_high: u16) -> Result<(), DispatchError> {
+        SubtensorModule::set_alpha_high(netuid, alpha_high)
+    }
+
+    fn set_alpha_low(netuid: u16, alpha_low: u16) -> Result<(), DispatchError> {
+        SubtensorModule::set_alpha_low(netuid, alpha_low)
+    }
+
+    fn set_liquid_alpha_enabled(netuid: u16, enabled: bool) {
+        SubtensorModule::set_liquid_alpha_enabled(netuid, enabled);
     }
 }
