+++ pallets/subtensor/src/subnet_info.rs
@@ -52,6 +52,9 @@ pub struct SubnetHyperparams {
     difficulty: Compact<u64>,
     commit_reveal_weights_interval: Compact<u64>,
     commit_reveal_weights_enabled: bool,
+    alpha_high: Compact<u16>,
+    alpha_low: Compact<u16>,
+    liquid_alpha_enabled: bool,
 }
 
 impl<T: Config> Pallet<T> {
@@ -155,6 +158,9 @@ impl<T: Config> Pallet<T> {
         let difficulty = Self::get_difficulty_as_u64(netuid);
         let commit_reveal_weights_interval = Self::get_commit_reveal_weights_interval(netuid);
         let commit_reveal_weights_enabled = Self::get_commit_reveal_weights_enabled(netuid);
+        let alpha_high = AlphaHigh::<T>::get(netuid);
+        let alpha_low = AlphaLow::<T>::get(netuid);
+        let liquid_alpha_enabled = LiquidAlphaOn::<T>::get(netuid);
 
         Some(SubnetHyperparams {
             rho: rho.into(),
@@ -181,6 +187,9 @@ impl<T: Config> Pallet<T> {
             difficulty: difficulty.into(),
             commit_reveal_weights_interval: commit_reveal_weights_interval.into(),
             commit_reveal_weights_enabled,
+            alpha_high: alpha_high.into(),
+            alpha_low: alpha_low.into(),
+            liquid_alpha_enabled,
         })
     }
 }
