+++ pallets/subtensor/src/utils.rs
@@ -1,6 +1,10 @@
 use super::*;
-use crate::system::{ensure_root, ensure_signed_or_root};
+use crate::{
+    system::{ensure_root, ensure_signed_or_root},
+    Error,
+};
 use sp_core::U256;
+use substrate_fixed::types::I32F32;
 
 impl<T: Config> Pallet<T> {
     pub fn ensure_subnet_owner_or_root(
@@ -657,5 +661,41 @@ impl<T: Config> Pallet<T> {
 
     pub fn set_nominator_min_required_stake(min_stake: u64) {
         NominatorMinRequiredStake::<T>::put(min_stake);
+    }
+
+    pub fn get_alpha_high(netuid: u16) -> I32F32 {
+        I32F32::from_num(AlphaHigh::<T>::get(netuid) as f64 / 1000.0)
+    }
+
+    pub fn set_alpha_high(netuid: u16, alpha_high: u16) -> Result<(), DispatchError> {
+        ensure!(
+            Self::get_liquid_alpha_enabled(netuid),
+            Error::<T>::LiquidAlphaDisabled
+        );
+        AlphaHigh::<T>::insert(netuid, alpha_high);
+
+        Ok(())
+    }
+
+    pub fn get_alpha_low(netuid: u16) -> I32F32 {
+        I32F32::from_num(AlphaLow::<T>::get(netuid) as f64 / 1000.0)
+    }
+
+    pub fn set_alpha_low(netuid: u16, alpha_low: u16) -> Result<(), DispatchError> {
+        ensure!(
+            Self::get_liquid_alpha_enabled(netuid),
+            Error::<T>::LiquidAlphaDisabled
+        );
+        AlphaLow::<T>::insert(netuid, alpha_low);
+
+        Ok(())
+    }
+
+    pub fn set_liquid_alpha_enabled(netuid: u16, enabled: bool) {
+        LiquidAlphaOn::<T>::set(netuid, enabled);
+    }
+
+    pub fn get_liquid_alpha_enabled(netuid: u16) -> bool {
+        LiquidAlphaOn::<T>::get(netuid)
     }
 }
