+++ pallets/subtensor/tests/epoch.rs
@@ -1,6 +1,7 @@
 use crate::mock::*;
-use frame_support::assert_ok;
+use frame_support::{assert_err, assert_ok};
 use frame_system::Config;
+use pallet_subtensor::*;
 use rand::{distributions::Uniform, rngs::StdRng, seq::SliceRandom, thread_rng, Rng, SeedableRng};
 use sp_core::U256;
 use std::time::Instant;
@@ -1262,6 +1263,322 @@ fn test_bonds() {
 	});
 }
 
+#[test]
+fn test_bonds_with_liquid_alpha() {
+    new_test_ext(1).execute_with(|| {
+        let sparse: bool = true;
+        let n: u16 = 8;
+        let netuid: u16 = 1;
+        let tempo: u16 = u16::MAX - 1; // high tempo to skip automatic epochs in on_initialize, use manual epochs instead
+        let max_stake: u64 = 4;
+        let stakes: Vec<u64> = vec![1, 2, 3, 4, 0, 0, 0, 0];
+        let block_number = System::block_number();
+        add_network(netuid, tempo, 0);
+        SubtensorModule::set_max_allowed_uids(netuid, n);
+        SubtensorModule::set_max_registrations_per_block(netuid, n);
+        SubtensorModule::set_target_registrations_per_interval(netuid, n);
+        SubtensorModule::set_weights_set_rate_limit(netuid, 0);
+        SubtensorModule::set_min_allowed_weights(netuid, 1);
+        SubtensorModule::set_max_weight_limit(netuid, u16::MAX);
+
+        // Register validators and servers
+        for key in 0..n as u64 {
+            SubtensorModule::add_balance_to_coldkey_account(&U256::from(key), max_stake);
+            let (nonce, work): (u64, Vec<u8>) = SubtensorModule::create_work_for_block_number(
+                netuid,
+                block_number,
+                key * 1_000_000,
+                &U256::from(key),
+            );
+            assert_ok!(SubtensorModule::register(
+                <<Test as Config>::RuntimeOrigin>::signed(U256::from(key)),
+                netuid,
+                block_number,
+                nonce,
+                work,
+                U256::from(key),
+                U256::from(key)
+            ));
+            SubtensorModule::increase_stake_on_coldkey_hotkey_account(
+                &U256::from(key),
+                &U256::from(key),
+                stakes[key as usize],
+            );
+        }
+
+        // Initilize with first epoch
+        SubtensorModule::epoch(netuid, 1_000_000_000);
+        next_block();
+
+        // Set weights
+        for uid in 0..(n / 2) {
+            SubtensorModule::set_validator_permit_for_uid(netuid, uid, true);
+            assert_ok!(SubtensorModule::set_weights(
+                RuntimeOrigin::signed(U256::from(uid)),
+                netuid,
+                ((n / 2)..n).collect(),
+                vec![u16::MAX / 4, u16::MAX / 2, (u16::MAX / 4) * 3, u16::MAX],
+                0
+            ));
+        }
+
+        // Enable Liquid Alpha
+        SubtensorModule::set_liquid_alpha_enabled(netuid, true);
+        assert_ok!(SubtensorModule::set_alpha_high(netuid, 900));
+        assert_ok!(SubtensorModule::set_alpha_low(netuid, 100));
+        // Run epoch with Liquid Alpha
+        if sparse {
+            SubtensorModule::epoch(netuid, 1_000_000_000);
+        } else {
+            SubtensorModule::epoch_dense(netuid, 1_000_000_000);
+        }
+
+        // Check bonds and emissions
+        let bonds = SubtensorModule::get_bonds(netuid);
+
+        /*  n: 8
+            current_block: 2; activity_cutoff: 5000;
+            Last update: [1, 1, 1, 1, 0, 0, 0, 0]
+            activity_cutoff: 5000
+            Last update: [2, 2, 2, 2, 1, 1, 1, 1]
+            Inactive: [false, false, false, false, false, false, false, false]
+            Block at registration: [1, 1, 1, 1, 1, 1, 1, 1]
+            hotkeys: [(0, 0), (1, 1), (2, 2), (3, 3), (4, 4), (5, 5), (6, 6), (7, 7)]
+            Stake: [1, 2, 3, 4, 0, 0, 0, 0]
+            Normalised Stake: [0.0999999999, 0.2, 0.2999999998, 0.4, 0, 0, 0, 0]
+            validator_permits: [true, true, true, true, true, true, true, true]
+            max_allowed_validators: 8
+            new_validator_permits: [true, true, true, true, true, true, true, true]
+            Active Stake: [0.0999999999, 0.2, 0.2999999998, 0.4, 0, 0, 0, 0]
+            Weights: [[(4, 16383), (5, 32767), (6, 49149), (7, 65535)], [(4, 16383), (5, 32767), (6, 49149), (7, 65535)], [(4, 16383), (5, 32767), (6, 49149), (7, 65535)], [(4, 16383), (5, 32767), (6, 49149), (7, 65535)], [], [], [], []]
+            Weights (permit): [[(4, 16383), (5, 32767), (6, 49149), (7, 65535)], [(4, 16383), (5, 32767), (6, 49149), (7, 65535)], [(4, 16383), (5, 32767), (6, 49149), (7, 65535)], [(4, 16383), (5, 32767), (6, 49149), (7, 65535)], [], [], [], []]
+            Weights (permit+diag): [[(4, 16383), (5, 32767), (6, 49149), (7, 65535)], [(4, 16383), (5, 32767), (6, 49149), (7, 65535)], [(4, 16383), (5, 32767), (6, 49149), (7, 65535)], [(4, 16383), (5, 32767), (6, 49149), (7, 65535)], [], [], [], []]
+            Weights (permit+diag+outdate): [[(4, 16383), (5, 32767), (6, 49149), (7, 65535)], [(4, 16383), (5, 32767), (6, 49149), (7, 65535)], [(4, 16383), (5, 32767), (6, 49149), (7, 65535)], [(4, 16383), (5, 32767), (6, 49149), (7, 65535)], [], [], [], []]
+            Weights (mask+norm): [[(4, 0.0999975584), (5, 0.2000012207), (6, 0.2999926754), (7, 0.400008545)], [(4, 0.0999975584), (5, 0.2000012207), (6, 0.2999926754), (7, 0.400008545)], [(4, 0.0999975584), (5, 0.2000012207), (6, 0.2999926754), (7, 0.400008545)], [(4, 0.0999975584), (5, 0.2000012207), (6, 0.2999926754), (7, 0.400008545)], [], [], [], []]
+            Ranks (before): [0, 0, 0, 0, 0.099997558, 0.2000012202, 0.2999926745, 0.4000085443]
+            Consensus: [0, 0, 0, 0, 0.0999975584, 0.2000012207, 0.2999926754, 0.400008545]
+            Weights: [[(4, 0.0999975584), (5, 0.2000012207), (6, 0.2999926754), (7, 0.400008545)], [(4, 0.0999975584), (5, 0.2000012207), (6, 0.2999926754), (7, 0.400008545)], [(4, 0.0999975584), (5, 0.2000012207), (6, 0.2999926754), (7, 0.400008545)], [(4, 0.0999975584), (5, 0.2000012207), (6, 0.2999926754), (7, 0.400008545)], [], [], [], []]
+            Validator Trust: [0.9999999995, 0.9999999995, 0.9999999995, 0.9999999995, 0, 0, 0, 0]
+            Ranks (after): [0, 0, 0, 0, 0.099997558, 0.2000012202, 0.2999926745, 0.4000085443]
+            T: [0, 0, 0, 0, 1, 1, 1, 1]
+            Incentive (=Rank): [0, 0, 0, 0, 0.0999975582, 0.2000012207,            0, 0.0999975582, 0.2000012207, 0.2999926752, 0.4000085455]
+            B: [[], [], [], [], [], [], [], []]
+            B (outdatedmask): [[], [], [], [], [], [], [], []]
+            B (mask+norm): [[], [], [], [], [], [], [], []]
+            ΔB: [[(4, 0.0099997558), (5, 0.020000122), (6, 0.0299992673), (7, 0.0400008543)], [(4, 0.0199995115), (5, 0.040000244), (6, 0.0599985349), (7, 0.0800017088)], [(4, 0.0299992673), (5, 0.060000366), (6, 0.0899978024), (7, 0.1200025633)], [(4, 0.0399990233), (5, 0.080000488), (6, 0.11999707), (7, 0.1600034179)], [], [], [], []]
+            ΔB (norm): [[(4, 0.0999999996), (5, 0.0999999999), (6, 0.0999999994), (7, 0.0999999996)], [(4, 0.1999999995), (5, 0.2), (6, 0.1999999997), (7, 0.1999999997)], [(4, 0.299999999), (5, 0.2999999998), (6, 0.3), (7, 0.3)], [(4, 0.4000000013), (5, 0.4), (6, 0.4000000004), (7, 0.4000000001)], [], [], [], []]
+            Exponential Moving Average Bonds Liquid Alpha: [[(4, 0.0499983232), (5, 0.0899999999), (6, 0.0899999994), (7, 0.0899999996)], [(4, 0.0999966469), (5, 0.18), (6, 0.1799999997), (7, 0.1799999997)], [(4, 0.1499949703), (5, 0.2699999998), (6, 0.2699999998), (7, 0.2699999998)], [(4, 0.199993295), (5, 0.3599999999), (6, 0.36), (7, 0.3599999999)], [], [], [], []]
+            Exponential Moving Average Bonds: [[(4, 0.0999999992), (5, 0.0999999999), (6, 0.0999999994), (7, 0.0999999996)], [(4, 0.1999999995), (5, 0.2), (6, 0.1999999997), (7, 0.1999999997)], [(4, 0.2999999993), (5, 0.2999999998), (6, 0.3), (7, 0.3)], [(4, 0.4000000015), (5, 0.4), (6, 0.4000000004), (7, 0.4000000001)], [], [], [], []]
+            Dividends: [0.0999999994, 0.1999999997, 0.3, 0.4000000006, 0, 0, 0, 0]
+            Normalized Server Emission: [0, 0, 0, 0, 0.049998779, 0.1000006103, 0.1499963375, 0.2000042726]
+            Server Emission: [0, 0, 0, 0, 49998779, 100000610, 149996337, 200004272]
+            Normalized Validator Emission: [0.0499999996, 0.0999999999, 0.15, 0.2000000002, 0, 0, 0, 0]
+            Validator Emission: [49999999, 99999999, 149999999, 200000000, 0, 0, 0, 0]
+            Normalized Combined Emission: [0.0499999996, 0.0999999999, 0.15, 0.2000000002, 0.049998779, 0.1000006103, 0.1499963375, 0.2000042726]
+            Combined Emission: [49999999, 99999999, 149999999, 200000000, 49998779, 100000610, 149996337, 200004272]
+            Pruning Scores: [0.0499999996, 0.0999999999, 0.15, 0.2000000002, 0.049998779, 0.1000006103, 0.1499963375, 0.2000042726]
+        */
+
+        // Expected bonds calculations
+        // For uid 0:
+        // Initial weights: [0.25, 0.5, 0.75, 1.0]
+        // Active stake: [1, 2, 3, 4]
+        // ΔB = W◦S = [0.25*1, 0.5*2, 0.75*3, 1.0*4] = [0.25, 1.0, 2.25, 4.0]
+        // Normalize ΔB: [0.25/7.5, 1.0/7.5, 2.25/7.5, 4.0/7.5] = [0.0333, 0.1333, 0.3, 0.5333]
+        // Final bonds for netuid: [16383, 32767, 49151, 65535]
+
+        assert_eq!(bonds[0][4], 16383); // Note: Calculated as explained above
+        assert_eq!(bonds[1][4], 32767); // Note: Calculated as explained above
+        assert_eq!(bonds[2][4], 49151); // Note: Calculated as explained above
+        assert_eq!(bonds[3][4], 65535); // Note: Calculated as explained above
+
+        // === Set self-weight only on val1
+        let uid = 0;
+        assert_ok!(SubtensorModule::set_weights(
+            RuntimeOrigin::signed(U256::from(uid)),
+            netuid,
+            vec![uid],
+            vec![u16::MAX],
+            0
+        ));
+        next_block();
+        if sparse {
+            SubtensorModule::epoch(netuid, 1_000_000_000);
+        } else {
+            SubtensorModule::epoch_dense(netuid, 1_000_000_000);
+        }
+
+        let bonds = SubtensorModule::get_bonds(netuid);
+        assert_eq!(bonds[0][4], 14582);
+        assert_eq!(bonds[1][4], 32767);
+        assert_eq!(bonds[2][4], 49151);
+        assert_eq!(bonds[3][4], 65535);
+
+        // === Set self-weight only on val2
+        let uid = 1;
+        assert_ok!(SubtensorModule::set_weights(
+            RuntimeOrigin::signed(U256::from(uid)),
+            netuid,
+            vec![uid],
+            vec![u16::MAX],
+            0
+        ));
+        next_block();
+        if sparse {
+            SubtensorModule::epoch(netuid, 1_000_000_000);
+        } else {
+            SubtensorModule::epoch_dense(netuid, 1_000_000_000);
+        }
+        let bonds = SubtensorModule::get_bonds(netuid);
+
+        /*  n: 8
+            current_block: 4; activity_cutoff: 5000;
+            Last update: [2, 3, 2, 2, 1, 1, 1, 1]
+            Inactive: [false, false, false, false, false, false, false, false]
+            Block at registration: [1, 1, 1, 1, 1, 1, 1, 1]
+            hotkeys: [(0, 0), (1, 1), (2, 2), (3, 3), (4, 4), (5, 5), (6, 6), (7, 7)]
+            Stake: [1, 2, 3, 4, 0, 0, 0, 0]
+            Normalised Stake: [0.0999999999, 0.2, 0.2999999998, 0.4, 0, 0, 0, 0]
+            validator_permits: [true, true, true, true, true, true, true, true]
+            max_allowed_validators: 64
+            new_validator_permits: [true, true, true, true, true, true, true, true]
+            Active Stake: [0.0999999999, 0.2, 0.2999999998, 0.4, 0, 0, 0, 0]
+            Weights: [[(0, 65535)], [(1, 65535)], [(4, 16383), (5, 32767), (6, 49149), (7, 65535)], [(4, 16383), (5, 32767), (6, 49149), (7, 65535)], [], [], [], []]
+            Weights (permit): [[(0, 65535)], [(1, 65535)], [(4, 16383), (5, 32767), (6, 49149), (7, 65535)], [(4, 16383), (5, 32767), (6, 49149), (7, 65535)], [], [], [], []]
+            Weights (permit+diag): [[], [], [(4, 16383), (5, 32767), (6, 49149), (7, 65535)], [(4, 16383), (5, 32767), (6, 49149), (7, 65535)], [], [], [], []]
+            Weights (permit+diag+outdate): [[], [], [(4, 16383), (5, 32767), (6, 49149), (7, 65535)], [(4, 16383), (5, 32767), (6, 49149), (7, 65535)], [], [], [], []]
+            Weights (mask+norm): [[], [], [(4, 0.0999975584), (5, 0.2000012207), (6, 0.2999926754), (7, 0.400008545)], [(4, 0.0999975584), (5, 0.2000012207), (6, 0.2999926754), (7, 0.400008545)], [], [], [], []]
+            Ranks (before): [0, 0, 0, 0, 0.0699982906, 0.1400008542, 0.2099948723, 0.2800059812]
+            Consensus: [0, 0, 0, 0, 0.0999975584, 0.2000012207, 0.2999926754, 0.400008545]
+            Weights: [[], [], [(4, 0.0999975584), (5, 0.2000012207), (6, 0.2999926754), (7, 0.400008545)], [(4, 0.0999975584), (5, 0.2000012207), (6, 0.2999926754), (7, 0.400008545)], [], [], [], []]
+            Validator Trust: [0, 0, 0.9999999995, 0.9999999995, 0, 0, 0, 0]
+            Ranks (after): [0, 0, 0, 0, 0.0699982906, 0.1400008542, 0.2099948723, 0.2800059812]
+            T: [0, 0, 0, 0, 1, 1, 1, 1]
+            Incentive (=Rank): [0, 0, 0, 0, 0.0999975582, 0.2000012207, 0.2999926754, 0.4000085455]
+            B: [[(4, 7760), (5, 1489), (6, 1489), (7, 1489)], [(4, 32767), (5, 32767), (6, 32767), (7, 32767)], [(4, 49151), (5, 49151), (6, 49151), (7, 49151)], [(4, 65535), (5, 65535), (6, 65535), (7, 65535)], [], [], [], []]
+            B (outdatedmask): [[(4, 7760), (5, 1489), (6, 1489), (7, 1489)], [(4, 32767), (5, 32767), (6, 32767), (7, 32767)], [(4, 49151), (5, 49151), (6, 49151), (7, 49151)], [(4, 65535), (5, 65535), (6, 65535), (7, 65535)], [], [], [], []]
+            B (mask+norm): [[(4, 0.0499958121), (5, 0.00999718), (6, 0.00999718), (7, 0.00999718)], [(4, 0.211109894), (5, 0.2199983886), (6, 0.2199983886), (7, 0.2199983886)], [(4, 0.3166680625), (5, 0.3300009398), (6, 0.3300009398), (7, 0.3300009398)], [(4, 0.4222262308), (5, 0.4400034912), (6, 0.4400034912), (7, 0.4400034912)], [], [], [], []]
+            ΔB: [[], [], [(4, 0.0299992673), (5, 0.060000366), (6, 0.0899978024), (7, 0.1200025633)], [(4, 0.0399990233), (5, 0.080000488), (6, 0.11999707), (7, 0.1600034179)], [], [], [], []]
+            ΔB (norm): [[], [], [(4, 0.428571427), (5, 0.4285714284), (6, 0.4285714284), (7, 0.4285714284)], [(4, 0.5714285728), (5, 0.5714285714), (6, 0.5714285714), (7, 0.5714285714)], [], [], [], []]
+            Exponential Moving Average Bonds Liquid Alpha: [[(4, 0.024998744), (5, 0.000999718), (6, 0.000999718), (7, 0.000999718)], [(4, 0.105558486), (5, 0.0219998388), (6, 0.0219998388), (7, 0.0219998388)], [(4, 0.3726178685), (5, 0.4187143792), (6, 0.4187143792), (7, 0.4187143792)], [(4, 0.4968249004), (5, 0.5582860631), (6, 0.5582860631), (7, 0.5582860631)], [], [], [], []]
+            Exponential Moving Average Bonds: [[(4, 0.024998744), (5, 0.000999718), (6, 0.000999718), (7, 0.000999718)], [(4, 0.105558486), (5, 0.0219998388), (6, 0.0219998388), (7, 0.0219998388)], [(4, 0.3726178687), (5, 0.4187143794), (6, 0.4187143794), (7, 0.4187143794)], [(4, 0.4968249009), (5, 0.5582860636), (6, 0.5582860636), (7, 0.5582860636)], [], [], [], []]
+            Dividends: [0.0033995616, 0.030355499, 0.4141048414, 0.5521400978, 0, 0, 0, 0]
+            Normalized Server Emission: [0, 0, 0, 0, 0.049998779, 0.1000006103, 0.1499963377, 0.2000042726]
+            Server Emission: [0, 0, 0, 0, 49998779, 100000610, 149996337, 200004272]
+            Normalized Validator Emission: [0.0016997808, 0.0151777493, 0.2070524206, 0.2760700488, 0, 0, 0, 0]
+            Validator Emission: [1699780, 15177749, 207052420, 276070048, 0, 0, 0, 0]
+            Normalized Combined Emission: [0.0016997808, 0.0151777493, 0.2070524206, 0.2760700488, 0.049998779, 0.1000006103, 0.1499963377, 0.2000042726]
+            Combined Emission: [1699780, 15177749, 207052420, 276070048, 49998779, 100000610, 149996337, 200004272]
+            Pruning Scores: [0.0016997808, 0.0151777493, 0.2070524206, 0.2760700488, 0.049998779, 0.1000006103, 0.1499963377, 0.2000042726]
+        */
+
+        assert_eq!(bonds[0][4], 12603);
+        assert_eq!(bonds[1][4], 28321);
+        assert_eq!(bonds[2][4], 49151);
+        assert_eq!(bonds[3][4], 65535);
+    });
+}
+
+#[test]
+fn test_bonds_with_extreme_alpha_values() {
+    new_test_ext(1).execute_with(|| {
+        let sparse: bool = true;
+        let n: u16 = 8;
+        let netuid: u16 = 1;
+        let tempo: u16 = u16::MAX - 1;
+        let max_stake: u64 = 4;
+        let stakes: Vec<u64> = vec![1, 2, 3, 4, 0, 0, 0, 0];
+        let block_number = System::block_number();
+        add_network(netuid, tempo, 0);
+        SubtensorModule::set_max_allowed_uids(netuid, n);
+        SubtensorModule::set_max_registrations_per_block(netuid, n);
+        SubtensorModule::set_target_registrations_per_interval(netuid, n);
+        SubtensorModule::set_weights_set_rate_limit(netuid, 0);
+        SubtensorModule::set_min_allowed_weights(netuid, 1);
+        SubtensorModule::set_max_weight_limit(netuid, u16::MAX);
+
+        for key in 0..n as u64 {
+            SubtensorModule::add_balance_to_coldkey_account(&U256::from(key), max_stake);
+            let (nonce, work) = SubtensorModule::create_work_for_block_number(
+                netuid,
+                block_number,
+                key * 1_000_000,
+                &U256::from(key),
+            );
+            assert_ok!(SubtensorModule::register(
+                RuntimeOrigin::signed(U256::from(key)),
+                netuid,
+                block_number,
+                nonce,
+                work,
+                U256::from(key),
+                U256::from(key)
+            ));
+            SubtensorModule::increase_stake_on_coldkey_hotkey_account(
+                &U256::from(key),
+                &U256::from(key),
+                stakes[key as usize],
+            );
+        }
+
+        SubtensorModule::epoch(netuid, 1_000_000_000);
+        next_block();
+
+        for uid in 0..(n / 2) {
+            SubtensorModule::set_validator_permit_for_uid(netuid, uid, true);
+            assert_ok!(SubtensorModule::set_weights(
+                RuntimeOrigin::signed(U256::from(uid)),
+                netuid,
+                ((n / 2)..n).collect(),
+                vec![u16::MAX / 4, u16::MAX / 2, (u16::MAX / 4) * 3, u16::MAX],
+                0
+            ));
+        }
+        SubtensorModule::set_liquid_alpha_enabled(netuid, true);
+        assert_ok!(SubtensorModule::set_alpha_high(netuid, u16::MAX));
+        assert_ok!(SubtensorModule::set_alpha_low(netuid, u16::MIN));
+        // Run epoch with Liquid Alpha
+        if sparse {
+            SubtensorModule::epoch(netuid, 1_000_000_000);
+        } else {
+            SubtensorModule::epoch_dense(netuid, 1_000_000_000);
+        }
+        let bonds = SubtensorModule::get_bonds(netuid);
+
+        log::info!("bonds: {:?}", bonds);
+
+        // Check for reasonable outputs despite extreme alpha values
+        assert!(bonds
+            .iter()
+            .flatten()
+            .all(|&bond| (0..=65535).contains(&bond)));
+    });
+}
+
+#[test]
+fn test_set_alpha_disabled() {
+    new_test_ext(1).execute_with(|| {
+        let netuid: u16 = 1;
+        let tempo: u16 = u16::MAX - 1;
+        add_network(netuid, tempo, 0);
+
+        // Explicitly set to false
+        SubtensorModule::set_liquid_alpha_enabled(netuid, false);
+        assert_err!(
+            SubtensorModule::set_alpha_high(netuid, u16::MAX),
+            Error::<Test>::LiquidAlphaDisabled
+        );
+        assert_err!(
+            SubtensorModule::set_alpha_low(netuid, 12_u16),
+            Error::<Test>::LiquidAlphaDisabled
+        );
+
+        SubtensorModule::set_liquid_alpha_enabled(netuid, true);
+        assert_ok!(SubtensorModule::set_alpha_high(netuid, u16::MAX));
+        assert_ok!(SubtensorModule::set_alpha_low(netuid, 12_u16));
+    });
+}
+
 // Test that epoch masks out inactive stake of validators with outdated weights beyond activity cutoff.
 #[test]
 fn test_active_stake() {
@@ -1972,6 +2289,385 @@ fn test_validator_permits() {
     }
 }
 
+#[test]
+fn test_compute_alpha_values() {
+    // Define the consensus values.
+    let consensus = vec![
+        I32F32::from_num(0.1),
+        I32F32::from_num(0.5),
+        I32F32::from_num(0.9),
+    ];
+    // Define the logistic function parameters 'a' and 'b'.
+    let a = I32F32::from_num(1.0);
+    let b = I32F32::from_num(0.0);
+
+    // Compute the alpha values using the function.
+    let alpha = SubtensorModule::compute_alpha_values(&consensus, a, b);
+
+    // Ensure the length of the alpha vector matches the consensus vector.
+    assert_eq!(alpha.len(), consensus.len());
+
+    // Manually compute the expected alpha values for each consensus value.
+    // The logistic function is: 1 / (1 + exp(b - a * c))
+    // where c is the consensus value.
+
+    // For consensus[0] = 0.1:
+    // exp_val = exp(0.0 - 1.0 * 0.1) = exp(-0.1)
+    // alpha[0] = 1 / (1 + exp(-0.1)) ~ 0.9048374180359595
+    let exp_val_0 = I32F32::from_num(0.9048374180359595);
+    let expected_alpha_0 =
+        I32F32::from_num(1.0).saturating_div(I32F32::from_num(1.0).saturating_add(exp_val_0));
+
+    // For consensus[1] = 0.5:
+    // exp_val = exp(0.0 - 1.0 * 0.5) = exp(-0.5)
+    // alpha[1] = 1 / (1 + exp(-0.5)) ~ 0.6065306597126334
+    let exp_val_1 = I32F32::from_num(0.6065306597126334);
+    let expected_alpha_1 =
+        I32F32::from_num(1.0).saturating_div(I32F32::from_num(1.0).saturating_add(exp_val_1));
+
+    // For consensus[2] = 0.9:
+    // exp_val = exp(0.0 - 1.0 * 0.9) = exp(-0.9)
+    // alpha[2] = 1 / (1 + exp(-0.9)) ~ 0.4065696597405991
+    let exp_val_2 = I32F32::from_num(0.4065696597405991);
+    let expected_alpha_2 =
+        I32F32::from_num(1.0).saturating_div(I32F32::from_num(1.0).saturating_add(exp_val_2));
+
+    // Define an epsilon for approximate equality checks.
+    let epsilon = I32F32::from_num(1e-6);
+
+    // Assert that the computed alpha values match the expected values within the epsilon.
+    assert_approx_eq(alpha[0], expected_alpha_0, epsilon);
+    assert_approx_eq(alpha[1], expected_alpha_1, epsilon);
+    assert_approx_eq(alpha[2], expected_alpha_2, epsilon);
+}
+
+#[test]
+fn test_clamp_alpha_values() {
+    // Define the alpha values.
+    let alpha = vec![
+        I32F32::from_num(0.1),
+        I32F32::from_num(0.5),
+        I32F32::from_num(0.9),
+    ];
+    // Define the high and low clamping values.
+    let alpha_high = I32F32::from_num(0.8);
+    let alpha_low = I32F32::from_num(0.2);
+
+    // Compute the clamped alpha values using the function.
+    let clamped_alpha = SubtensorModule::clamp_alpha_values(alpha.clone(), alpha_high, alpha_low);
+
+    // Ensure the length of the clamped alpha vector matches the original alpha vector.
+    assert_eq!(clamped_alpha.len(), alpha.len());
+
+    // Manually compute the expected clamped alpha values for each alpha value.
+    // The clamping logic is: max(alpha_low, min(alpha_high, a))
+
+    // For alpha[0] = 0.1:
+    // clamped_a = max(0.2, min(0.8, 0.1)) = max(0.2, 0.1) = 0.2
+    let expected_clamped_alpha_0 = I32F32::from_num(0.2);
+
+    // For alpha[1] = 0.5:
+    // clamped_a = max(0.2, min(0.8, 0.5)) = max(0.2, 0.5) = 0.5
+    let expected_clamped_alpha_1 = I32F32::from_num(0.5);
+
+    // For alpha[2] = 0.9:
+    // clamped_a = max(0.2, min(0.8, 0.9)) = max(0.2, 0.8) = 0.8
+    let expected_clamped_alpha_2 = I32F32::from_num(0.8);
+
+    // Assert that the computed clamped alpha values match the expected values.
+    assert_eq!(clamped_alpha[0], expected_clamped_alpha_0);
+    assert_eq!(clamped_alpha[1], expected_clamped_alpha_1);
+    assert_eq!(clamped_alpha[2], expected_clamped_alpha_2);
+}
+
+#[test]
+fn test_calculate_logistic_params() {
+    // Define test inputs
+    let alpha_high = I32F32::from_num(0.9);
+    let alpha_low = I32F32::from_num(0.1);
+    let consensus_high = I32F32::from_num(0.8);
+    let consensus_low = I32F32::from_num(0.2);
+
+    // Expected values
+    // a = (ln((1 / alpha_high - 1)) - ln((1 / alpha_low - 1))) / (consensus_low - consensus_high)
+    //   = (ln((1 / 0.9 - 1)) - ln((1 / 0.1 - 1))) / (0.2 - 0.8)
+    //   = (ln(0.1111) - ln(9)) / -0.6
+    //   = (-2.1972 - 2.1972) / -0.6
+    //   = -4.3944 / -0.6
+    //   = 7.324
+    let expected_a = I32F32::from_num(7.324);
+
+    // b = ln((1 / alpha_low - 1)) + a * consensus_low
+    //   = ln((1 / 0.1 - 1)) + 7.324 * 0.2
+    //   = ln(9) + 1.4648
+    //   = 2.1972 + 1.4648
+    //   = 3.662
+    let expected_b = I32F32::from_num(3.662);
+
+    // Call the function
+    let (a, b) = SubtensorModule::calculate_logistic_params(
+        alpha_high,
+        alpha_low,
+        consensus_high,
+        consensus_low,
+    );
+
+    // Assert the results
+    assert!(
+        (a - expected_a).abs() < I32F32::from_num(0.001),
+        "Expected a: {:?}, got: {:?}",
+        expected_a,
+        a
+    );
+    assert!(
+        (b - expected_b).abs() < I32F32::from_num(0.001),
+        "Expected b: {:?}, got: {:?}",
+        expected_b,
+        b
+    );
+}
+
+#[test]
+fn test_calculate_logistic_params_edge_cases() {
+    // Edge Case 1: Alpha values at their boundaries (0 and 1)
+    let alpha_high = I32F32::from_num(1.0);
+    let alpha_low = I32F32::from_num(0.0);
+    let consensus_high = I32F32::from_num(0.8);
+    let consensus_low = I32F32::from_num(0.2);
+
+    // Call the function
+    let (a, b) = SubtensorModule::calculate_logistic_params(
+        alpha_high,
+        alpha_low,
+        consensus_high,
+        consensus_low,
+    );
+
+    // Assert the results
+    assert_eq!(a, I32F32::from_num(0.0), "Expected a to be 0, got: {:?}", a);
+    assert_eq!(b, I32F32::from_num(0.0), "Expected b to be 0, got: {:?}", b);
+
+    // Edge Case 2: Consensus values at their boundaries (0 and 1)
+    let alpha_high = I32F32::from_num(0.9);
+    let alpha_low = I32F32::from_num(0.1);
+    let consensus_high = I32F32::from_num(1.0);
+    let consensus_low = I32F32::from_num(0.0);
+
+    // Call the function
+    let (a, b) = SubtensorModule::calculate_logistic_params(
+        alpha_high,
+        alpha_low,
+        consensus_high,
+        consensus_low,
+    );
+
+    // Expected values
+    // a = (ln((1 / 0.9 - 1)) - ln((1 / 0.1 - 1))) / (0.0 - 1.0)
+    //   = (ln(0.1111) - ln(9)) / -1.0
+    //   = (-2.1972 - 2.1972) / -1.0
+    //   = -4.3944 / -1.0
+    //   = 4.3944
+    let expected_a = I32F32::from_num(4.3944);
+
+    // b = ln((1 / 0.1 - 1)) + a * 0.0
+    //   = ln(9) + 0
+    //   = 2.1972
+    let expected_b = I32F32::from_num(2.1972);
+
+    // Assert the results
+    assert!(
+        (a - expected_a).abs() < I32F32::from_num(0.001),
+        "Expected a: {:?}, got: {:?}",
+        expected_a,
+        a
+    );
+    assert!(
+        (b - expected_b).abs() < I32F32::from_num(0.001),
+        "Expected b: {:?}, got: {:?}",
+        expected_b,
+        b
+    );
+
+    // Edge Case 3: Alpha values being equal
+    let alpha_high = I32F32::from_num(0.5);
+    let alpha_low = I32F32::from_num(0.5);
+    let consensus_high = I32F32::from_num(0.8);
+    let consensus_low = I32F32::from_num(0.2);
+
+    // Call the function
+    let (a, b) = SubtensorModule::calculate_logistic_params(
+        alpha_high,
+        alpha_low,
+        consensus_high,
+        consensus_low,
+    );
+
+    // Assert the results
+    assert_eq!(a, I32F32::from_num(0.0), "Expected a to be 0, got: {:?}", a);
+    assert_eq!(b, I32F32::from_num(0.0), "Expected b to be 0, got: {:?}", b);
+
+    // Edge Case 4: Consensus values being equal
+    let alpha_high = I32F32::from_num(0.9);
+    let alpha_low = I32F32::from_num(0.1);
+    let consensus_high = I32F32::from_num(0.5);
+    let consensus_low = I32F32::from_num(0.5);
+
+    // Call the function
+    let (a, b) = SubtensorModule::calculate_logistic_params(
+        alpha_high,
+        alpha_low,
+        consensus_high,
+        consensus_low,
+    );
+
+    // Assert the results
+    assert_eq!(a, I32F32::from_num(0.0), "Expected a to be 0, got: {:?}", a);
+    assert_eq!(b, I32F32::from_num(0.0), "Expected b to be 0, got: {:?}", b);
+}
+
+#[test]
+fn test_compute_ema_bonds_with_liquid_alpha_sparse() {
+    // Define test inputs
+    let bonds_delta = vec![
+        vec![(0, I32F32::from_num(0.1)), (1, I32F32::from_num(0.2))],
+        vec![(0, I32F32::from_num(0.3)), (1, I32F32::from_num(0.4))],
+    ];
+    let bonds = vec![
+        vec![(0, I32F32::from_num(0.5)), (1, I32F32::from_num(0.6))],
+        vec![(0, I32F32::from_num(0.7)), (1, I32F32::from_num(0.8))],
+    ];
+    let alpha = vec![I32F32::from_num(0.9), I32F32::from_num(0.8)];
+
+    // Expected values
+    // EMA calculation for each bond:
+    // EMA = alpha * bond_delta + (1 - alpha) * bond
+    // For bond (0, 0):
+    // EMA = 0.9 * 0.1 + (1 - 0.9) * 0.5 = 0.09 + 0.05 = 0.14
+    // For bond (0, 1):
+    // EMA = 0.8 * 0.2 + (1 - 0.8) * 0.6 = 0.16 + 0.12 = 0.28
+    // For bond (1, 0):
+    // EMA = 0.9 * 0.3 + (1 - 0.9) * 0.7 = 0.27 + 0.07 = 0.34
+    // For bond (1, 1):
+    // EMA = 0.8 * 0.4 + (1 - 0.8) * 0.8 = 0.32 + 0.16 = 0.48
+    let expected_ema_bonds = vec![
+        vec![(0, I32F32::from_num(0.14)), (1, I32F32::from_num(0.28))],
+        vec![(0, I32F32::from_num(0.34)), (1, I32F32::from_num(0.48))],
+    ];
+
+    // Call the function
+    let ema_bonds =
+        SubtensorModule::compute_ema_bonds_with_liquid_alpha_sparse(&bonds_delta, &bonds, alpha);
+
+    // Assert the results with an epsilon for approximate equality
+    let epsilon = I32F32::from_num(1e-6);
+    assert_approx_eq_vec_of_vec(&ema_bonds, &expected_ema_bonds, epsilon);
+}
+
+#[test]
+fn test_compute_ema_bonds_with_liquid_alpha_sparse_empty() {
+    // Test with empty inputs
+    let bonds_delta: Vec<Vec<(u16, I32F32)>> = vec![];
+    let bonds: Vec<Vec<(u16, I32F32)>> = vec![];
+    let alpha: Vec<I32F32> = vec![];
+
+    // Expected values: Empty Vec
+    let expected_ema_bonds: Vec<Vec<(u16, I32F32)>> = vec![];
+
+    // Call the function
+    let ema_bonds =
+        SubtensorModule::compute_ema_bonds_with_liquid_alpha_sparse(&bonds_delta, &bonds, alpha);
+
+    // Assert the results
+    assert_eq!(
+        ema_bonds, expected_ema_bonds,
+        "Expected EMA bonds: {:?}, got: {:?}",
+        expected_ema_bonds, ema_bonds
+    );
+}
+
+// #[test]
+// fn test_compute_ema_bonds_sparse_with_liquid_alpha() {
+//     new_test_ext(1).execute_with(|| {
+//         let sparse: bool = true;
+//         let n: u16 = 8;
+//         let netuid: u16 = 1;
+//         let tempo: u16 = u16::MAX - 1;  // high tempo to skip automatic epochs in on_initialize, use manual epochs instead
+//         let max_stake: u64 = 4;
+//         let stakes: Vec<u64> = vec![1, 2, 3, 4, 0, 0, 0, 0];
+//         let block_number = System::block_number();
+//         add_network(netuid, tempo, 0);
+//         SubtensorModule::set_max_allowed_uids(netuid, n);
+//         SubtensorModule::set_max_registrations_per_block(netuid, n);
+//         SubtensorModule::set_target_registrations_per_interval(netuid, n);
+//         SubtensorModule::set_weights_set_rate_limit(netuid, 0);
+//         SubtensorModule::set_min_allowed_weights(netuid, 1);
+//         SubtensorModule::set_max_weight_limit(netuid, u16::MAX);
+
+//         // Register validators and servers
+//         for key in 0..n as u64 {
+//             SubtensorModule::add_balance_to_coldkey_account(&U256::from(key), max_stake);
+//             let (nonce, work): (u64, Vec<u8>) = SubtensorModule::create_work_for_block_number(
+//                 netuid,
+//                 block_number,
+//                 key * 1_000_000,
+//                 &U256::from(key),
+//             );
+//             assert_ok!(SubtensorModule::register(
+//                 <<Test as Config>::RuntimeOrigin>::signed(U256::from(key)),
+//                 netuid,
+//                 block_number,
+//                 nonce,
+//                 work,
+//                 U256::from(key),
+//                 U256::from(key)
+//             ));
+//             SubtensorModule::increase_stake_on_coldkey_hotkey_account(
+//                 &U256::from(key),
+//                 &U256::from(key),
+//                 stakes[key as usize],
+//             );
+//         }
+
+//         // Initialize with first epoch
+//         SubtensorModule::epoch(netuid, 1_000_000_000);
+//         step_block(1);
+
+//         // Set weights
+//         for uid in 0..(n / 2) as u16 {
+//             SubtensorModule::set_validator_permit_for_uid(netuid, uid, true);
+//             assert_ok!(SubtensorModule::set_weights(
+//                 RuntimeOrigin::signed(U256::from(uid)),
+//                 netuid,
+//                 ((n / 2)..n).collect(),
+//                 vec![u16::MAX / 4, u16::MAX / 2, (u16::MAX / 4) * 3, u16::MAX],
+//                 0
+//             ));
+//         }
+
+//         // Enable LiquidAlpha
+//         SubtensorModule::set_liquid_alpha_enabled(netuid, true);
+//         assert_eq!(SubtensorModule::get_liquid_alpha_enabled(netuid), true);
+
+//         // Continue with additional epochs to mimic the end-to-end test
+//         // for _ in 0..5 {
+//         //     next_block();
+//         //     if sparse {
+//         //         SubtensorModule::epoch(netuid, 1_000_000_000);
+//         //     } else {
+//         //         SubtensorModule::epoch_dense(netuid, 1_000_000_000);
+//         //     }
+//         // }
+
+//         // Fetch the final bonds and validate
+//         let final_bonds = SubtensorModule::get_bonds(netuid);
+//         // Example assertions, adjust as needed based on expected LiquidAlpha values
+//         assert_eq!(final_bonds[0][0], 14582);
+//         assert_eq!(final_bonds[1][1], 32767);
+//         assert_eq!(final_bonds[2][2], 49151);
+//         assert_eq!(final_bonds[3][3], 65535);
+//     });
+// }
+
 // // Map the retention graph for consensus guarantees with an single epoch on a graph with 512 nodes, of which the first 64 are validators, the graph is split into a major and minor set, each setting specific weight on itself and the complement on the other.
 // //
 // // ```import torch
@@ -2070,3 +2766,47 @@ fn test_validator_permits() {
 //     }
 //     println!("]");
 // }
+
+/// Helpers
+
+/// Asserts that two I32F32 values are approximately equal within a given epsilon.
+///
+/// # Arguments
+/// * `left` - The first value to compare.
+/// * `right` - The second value to compare.
+/// * `epsilon` - The maximum allowed difference between the two values.
+fn assert_approx_eq(left: I32F32, right: I32F32, epsilon: I32F32) {
+    if (left - right).abs() > epsilon {
+        panic!(
+            "assertion failed: `(left ≈ right)`\n  left: `{:?}`,\n right: `{:?}`,\n epsilon: `{:?}`",
+            left, right, epsilon
+        );
+    }
+}
+
+/// Helper function to assert approximate equality of two vectors of vectors of tuples.
+fn assert_approx_eq_vec_of_vec(
+    left: &Vec<Vec<(u16, I32F32)>>,
+    right: &Vec<Vec<(u16, I32F32)>>,
+    epsilon: I32F32,
+) {
+    assert_eq!(left.len(), right.len(), "Vectors have different lengths");
+    for (left_row, right_row) in left.iter().zip(right.iter()) {
+        assert_eq!(
+            left_row.len(),
+            right_row.len(),
+            "Rows have different lengths"
+        );
+        for ((left_idx, left_val), (right_idx, right_val)) in left_row.iter().zip(right_row.iter())
+        {
+            assert_eq!(left_idx, right_idx, "Indices are different");
+            assert!(
+                (left_val - right_val).abs() < epsilon,
+                "Values are different: left = {:?}, right = {:?}, epsilon = {:?}",
+                left_val,
+                right_val,
+                epsilon
+            );
+        }
+    }
+}
